[IMP] Added tests_rest_users
