+++ tests/tests_rest.py
@@ -138,71 +138,6 @@ class TestsRest(TestCase):
         response = self._subject.create('/manage/settings/update', body)
         self.assertEqual(200, response.status_code)
 
-    def test_create_ioc_should_return_good_ioc_type_id(self):
-        case_identifier = self._subject.create_dummy_case()
-        body = {'ioc_type_id': 1, 'ioc_tlp_id': 2, 'ioc_value': '8.8.8.8', 'ioc_description': 'rewrw', 'ioc_tags': ''}
-        response = self._subject.create(f'/api/v2/cases/{case_identifier}/iocs', body).json()
-        self.assertEqual(1, response['ioc_type_id'])
-
-    def test_get_ioc_should_return_ioc_type_id(self):
-        ioc_type_id = 1
-        case_identifier = self._subject.create_dummy_case()
-        body = {'ioc_type_id': ioc_type_id, 'ioc_tlp_id': 2, 'ioc_value': '8.8.8.8', 'ioc_description': 'rewrw', 'ioc_tags': ''}
-        test = self._subject.create(f'/api/v2/cases/{case_identifier}/iocs', body).json()
-        current_id = test['ioc_id']
-        response = self._subject.get(f'/api/v2/iocs/{current_id}').json()
-        self.assertEqual(ioc_type_id, response['ioc_type_id'])
-
-    def test_get_ioc_with_missing_ioc_identifier_should_return_error(self):
-        case_identifier = self._subject.create_dummy_case()
-        body = {'ioc_type_id': 1, 'ioc_tlp_id': 2, 'ioc_value': '8.8.8.8', 'ioc_description': 'rewrw', 'ioc_tags': ''}
-        self._subject.create(f'/api/v2/cases/{case_identifier}/iocs', body).json()
-        test = self._subject.get('/api/v2/iocs/None').json()
-        self.assertEqual('error', test['status'])
-
-    def test_delete_ioc_should_return_204(self):
-        case_identifier = self._subject.create_dummy_case()
-        body = {'ioc_type_id': 1, 'ioc_tlp_id': 2, 'ioc_value': '8.8.8.8', 'ioc_description': 'rewrw', 'ioc_tags': ''}
-        response = self._subject.create(f'/api/v2/cases/{case_identifier}/iocs', body).json()
-        ioc_identifier = response['ioc_id']
-        response = self._subject.delete(f'/api/v2/iocs/{ioc_identifier}')
-        self.assertEqual(204, response.status_code)
-
-    def test_delete_ioc_with_missing_ioc_identifier_should_return_404(self):
-        response = self._subject.delete(f'/api/v2/iocs/{_IDENTIFIER_FOR_NONEXISTENT_OBJECT}')
-        self.assertEqual(404, response.status_code)
-
-    def test_delete_asset_should_return_204(self):
-        case_identifier = self._subject.create_dummy_case()
-        body = {'asset_type_id': '1', 'asset_name': 'admin_laptop_test'}
-        response = self._subject.create(f'/api/v2/cases/{case_identifier}/assets', body).json()
-        asset_identifier = response['asset_id']
-        response = self._subject.delete(f'/api/v2/assets/{asset_identifier}')
-        self.assertEqual(204, response.status_code)
-
-    def test_delete_asset_with_missing_asset_identifier_should_return_404(self):
-        response = self._subject.delete(f'/api/v2/assets/{_IDENTIFIER_FOR_NONEXISTENT_OBJECT}')
-        self.assertEqual(404, response.status_code)
-
-    def test_create_asset_should_work(self):
-        case_identifier = self._subject.create_dummy_case()
-        body = {'asset_type_id': '1', 'asset_name': 'admin_laptop_test'}
-        response = self._subject.create(f'/api/v2/cases/{case_identifier}/assets', body)
-        self.assertEqual(201, response.status_code)
-
-    def test_get_asset_with_missing_asset_identifier_should_return_404(self):
-        response = self._subject.get('/api/v2/asset/None')
-        self.assertEqual(404, response.status_code)
-
     def test_get_timeline_state_should_return_200(self):
         response = self._subject.get('/case/timeline/state', query_parameters={'cid': 1})
         self.assertEqual(200, response.status_code)
-
-    def test_get_users_should_return_200(self):
-        response = self._subject.get('/manage/users/list')
-        self.assertEqual(200, response.status_code)
-
-    def test_get_users_should_return_403_for_user_without_rights(self):
-        user = self._subject.create_dummy_user()
-        response = user.get('/manage/users/list')
-        self.assertEqual(403, response.status_code)
